backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
PyGithub>=2.1.1
pinecone>=5.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
tenacity>=8.2.0
